Move shell-state regression suite to its own Swift Testing file (Aziz policy)
cmuxTests convention (and the Aziz test-framework policy) is Swift Testing for
non-UI tests. Move the #6618 regression coverage out of the XCTest-based
TabManagerUnitTests.swift into a dedicated TabManagerShellActivityReplayTests.swift
Swift Testing suite (wired into the cmuxTests target), and document the
WorkspacesHosting default-implementations extension for the DocC policy.
